refactor: move controls to seperate component
